+++ packages/client/src/client.ts
@@ -11,6 +11,7 @@ import {
   RuntimeContext,
   getObservableClient,
 } from "@polkadot-api/observable-client"
+import { Binary } from "@polkadot-api/substrate-bindings"
 import {
   SubstrateClient,
   createClient as createRawClient,
@@ -35,11 +36,11 @@ import {
 import { createConstantEntry } from "./constants"
 import { ChainDefinition } from "./descriptors"
 import { createEventEntry } from "./event"
+import { getNewBlocks$, getPinnedBlocks$ } from "./pinnedBlocks"
 import { createRuntimeCallEntry } from "./runtime-call"
 import { createStorageEntry } from "./storage"
 import { createTxEntry, submit, submit$ } from "./tx"
 import type { AnyApi, PolkadotClient } from "./types"
-import { Binary } from "@polkadot-api/substrate-bindings"
 import { createWatchEntries } from "./watch-entries"
 
 const createApi = <Unsafe extends true | false, D>(
@@ -279,6 +280,10 @@ export function createClient(provider: JsonRpcProvider): PolkadotClient {
   const { broadcastTx$ } = client
   return {
     getChainSpecData,
+
+    blocks$: getNewBlocks$(chainHead.pinnedBlocks$),
+
+    pinnedBlocks$: getPinnedBlocks$(chainHead.pinnedBlocks$),
 
     finalizedBlock$: chainHead.finalized$,
     getFinalizedBlock: () => firstValueFrom(chainHead.finalized$),