+++ packages/observable-client/src/index.ts
@@ -3,7 +3,6 @@ export { isBestOrFinalizedBlock } from "./chainHead/streams/block-operations"
 export {
   BlockNotPinnedError,
   BlockPrunedError,
-  NotBestBlockError,
   type AnalyzedBlock,
   type PinnedBlock,
   type BlockInfo,